Move the logic for url downloading into the Source_Url class
